+++ packages/laconic_postgresql/lib/src/postgresql_driver.dart
@@ -16,7 +16,7 @@ import 'package:postgres/postgres.dart';
 /// ));
 /// final db = Laconic(driver);
 /// ```
-class PostgresqlDriver implements LaconicDriver {
+class PostgresqlDriver implements DatabaseDriver {
   final PostgresqlConfig config;
   Pool? _pool;